Change group by hash function to wyhash Change random strategy to wyrand

+++ src/operator/end.rs
@@ -21,7 +21,7 @@ use super::AsyncOperator;
 #[derivative(Clone, Debug)]
 pub struct EndBlock<Out: ExchangeData, OperatorChain, IndexFn>
 where
-    IndexFn: KeyerFn<usize, Out>,
+    IndexFn: KeyerFn<u64, Out>,
     OperatorChain: Operator<Out>,
 {
     prev: OperatorChain,
@@ -39,7 +39,7 @@ where
 
 impl<Out: ExchangeData, OperatorChain, IndexFn> EndBlock<Out, OperatorChain, IndexFn>
 where
-    IndexFn: KeyerFn<usize, Out>,
+    IndexFn: KeyerFn<u64, Out>,
     OperatorChain: Operator<Out>,
 {
     pub(crate) fn new(
@@ -165,7 +165,7 @@ where
 impl<Out: ExchangeData, OperatorChain, IndexFn> Operator<()>
     for EndBlock<Out, OperatorChain, IndexFn>
 where
-    IndexFn: KeyerFn<usize, Out>,
+    IndexFn: KeyerFn<u64, Out>,
     OperatorChain: Operator<Out>,
 {
     fn setup(&mut self, metadata: ExecutionMetadata) {
